Warnings fixed
<==== Mercurial data ====>
Branch: meta

+++ common_libs/nx_sdk/src/nx/sdk/common_settings.cpp
@@ -29,7 +29,7 @@ void CommonSettings::clear()
 
 int CommonSettings::count() const
 {
-    return m_settings.size();
+    return (int) m_settings.size();
 }
 
 const char* CommonSettings::key(int i) const
